requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.24.3
botocore==1.27.3
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-on-heroku==1.1.2
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
